requirements.txt
fastapi
uvicorn

langchain
langchain-community
langchain-openai
langchain-text-splitters

faiss-cpu

pypdf
docx2txt

python-dotenv